Fixing the summary of RTCTransportStats (issue #72).
The new summary should make it clear that there is a 1:1 mapping between
an RTCTransportStats and an RTCDtlsTransport/RTCIceTransport pair.
Meaning, when RTCP muxing is *not* used, you get separate
RTCTransportStats for RTP and RTCP. Also adding links to the appropriate
terms, and to WebRTC (to describe RTCP muxing and bundling in more
detail).