Adding "codec type" and transportId to RTCCodecStats.
Fixes #109.

The codec type indicates if it's a media format the implementation is
prepared to encode or decode, and the transportId identifies what RTP
session it's being used for.

An implementation of getStats may want to return RTCCodecStats objects
for codecs not being actively used, in which case these fields are
necessary in order for those objects to provide meaningful information.