Adding "networkType" field to RTCIceCandidateStats.
Fixes #257.

Exposes the type of network interface a local ICE candidate is using.
Defined such that it will be present for srflx/relay candidates as well,
exposing the type of network interface used to communicate with the
STUN/TURN server.

RTCNetworkType is based on this enum:
http://wicg.github.io/netinfo/#dom-connectiontype

With the only differences being that "none" isn't applicable, since every
ICE candidate is using *some* network interface, and "mixed" isn't
applicable since every ICE candidate is tied to only one interface.