Generate appendix with flatten hierarchy of stats
close #577
The postProcess function  looks at the stats type listed in the table under the appendix with id summary, and for each dictionary listed for the said type looks up the IDL fragment, parses it, identifies inherited dictionaries, and lists all the members for each of those (not mentioning empty dictionaries).

Only non-obsolete stats types are included in the said appendix.