Don't export the informative RTCStats reproduction
The `RTCStats` dictionary is normatively defined in WebRTC and reproduced in
WebRTC Stats for ease of reference only. The IDL block is correctly flagged as
informative, but ReSpec still exports the IDL term, meaning that `{{RTCStats}}`
can either be interpreted to mean:
1. a normative reference to the `RTCStats` dictionary defined in WebRTC
2. an informative reference to the `RTCStats` dictionary reproductionin WebRTC
Stats.

The informative reproduction should not be exported. This update tells ReSpec
not to export `RTCStats`. Note that there is no way (at least none that I'm
aware of) to do that directly within the IDL block, so this moves the definition
to the prose that precedes the IDL block.